SERVER-24751 add collation parameter to shardCollection command
This parameter must be used to specify the simple collation
when the collection has a non-simple default collation.
Since the shard key is required to have the simple
collation, we now enforce that there is an index whose
prefix is the shard key and whose collation is the simple
collation.