SERVER-27362 Avoid extra logic while getting lock to destroy M/R cursor.
The mapReduce commands opens a ClientCursor to prevent the RangeDeleter
from deleting data from chunks that have been migrated away from the
shard. To destroy the ClientCursor at the end of the command, it needs
to take a lock on the collection the cursor is registered on. It should
not use AutoGetCollectionForRead to acquire this lock since that will
do unnecessary extra tasks, including throwing an exception if the
shard version no longer matches. Instead, it should simply use
AutoGetCollection to get the required lock, to ensure the ClientCursor
can be cleaned up correctly.