SERVER-28837 Do not attempt to subtract negative cursorTimeoutMillis.
An INT_MIN value for cursorTimeoutMillis can crash mongos when we try
to subtract that value (which gets cast to a Duration) from the
current time. Because a negative cursorTimeoutMillis is semantically
the same as a 0 value, this change ignores negative values when
performing the offending subtraction.

It would be reasonable to clamp cursorTimeoutMillis to 0 using input
validation, but I don't see a good way to do that. Because a Duration
cannot be an unsigned type, we have to keep cursorTimeoutMillis as an
int. Unfortunately, we don't really have a way restrict the values
that the user sets with the setParameter command except to enforce
that the value fits in the data typed used for the parameter.