SERVER-29305 `dbdirectclient::call` uses OpContext
Presently it directly calls `assembleResponse`, which is responsible
for a cyclic dependency.  This functionality is available through
the `ServiceContext` mechanism, and thus needn't directly use
`assembleResponse`.

The `ServiceEntryPoint` member of `ServiceContext` is set through
a setter, which isn't always called, so we initialize the global
entry through a factory function.