BF-5630 Deadlock in PeriodicRunnerAsio test
Under certain loads, it's possible for the waiting (main) thread
to .fastForward the mock timer while the incrementer thread is
still returning. This yields no active job to trigger which means
the main thread never gets its next notification to complete
the cv.wait().

Since the incrementer thread isn't looped by a real timer,
it never fires again to release the waiting thread and we wind
up with a deadlock.

By introducing a wait at the end of the loop, we can ensure the
job has completed and rescheduled itself before its next manual
fastForward'ing.