SERVER-31271 don't use nspr allocator for js threads for better ASAN
We can't use the nspr allocator to allocate threads in PR_CreateThread, because under asan
instrument the allocator so that asan can track the pointers correctly. This instrumentation
requires that pointers be deleted in the same thread that they were allocated in.
The threads created in PR_CreateThread are not always freed in the same thread
that they were created in. So, we use the standard allocator here.