SERVER-31726: Trigger an OplogManager flush directly after transaction completion.
Readers have the ability to wait for oplog writes to become visible. The
visibility can be updated when a transaction commits or rolls back. There is a
background thread that polls WiredTiger for the visibility and publishes it to
readers.

For efficiency reasons, this background thread waits on a condition variable and
only iterates when notified. Previously, this notification was done via a "change"
callback on the recovery unit that is registered when a transaction writes its
entry to the oplog. However, there were (error) code paths where a transaction
finishing could have an impact on visibility, but not yet reached the point where
the callback was registered.

This patch fixes that error case by having transactions that can impact visibility
directly trigger the background thread when they complete.