SERVER-31648 Perform sets/unsets to session tags atomically.
As described in the SERVER ticket, there is a window in
CmdIsMaster::run() during which there is no flag set to prevent a
session from being closed by an FCV update (even when the session is
one that should stay open through the upgrade). This change closes
that window.

As part of the fix, this change also replaces the replaceTags()
function with functions that can atomically update the session tags.