SERVER-34506 TLASIO test for isJustForContinuation
Adding an integration test for transport layer asio which uses a fail
point to induce single byte at a time reads and writes.  We use this,
along with a debug block in the future header, to ensure that the
continuation folding in futures is working properly with
opportunisticRead/Write chaining.

To test the other side of this, adding a new test command called "echo"
which returns the command object passed to it (to allow for a large
message body on the read and write path more easily).