SERVER-36546 Be more cautious with key comparision during cursor restore.
The new format unique indexes *do* include the record id in their
KeyStrings. This means that restoring to the same key with a new
record id *should* return false so that we will skip the key with the new
record-id. To enable this we compare the current key with the key
at the saved position. This comparision should be done only if the
saved key has a size greater than or equal to the key being compared.