SERVER-41412 Introduce a search_beta_auth suite.
This variant of search_beta turns on auth on the cluster,
including on the mongotmock, which supports the SCRAM-SHA-1
authentication mechanism.

Also configures the pre-existing search_beta_ssl suite to
use auth.

There are changes in this commit which permit mongotmock to
link in the necessary auth-related code and command line
parameters without pulling in unnecessary dependencies, or
adding new dependencies to mongocryptd.