SERVER-43081 Validate should report when an index's 'multikeyPaths' are set but the 'multikey' flag is false
This also removes the invariant added by SERVER-43008 that performed the
same check in the catalog layer. It is not safe to invariant based on
durable data if it is known that data could be corrupt in an undefined
way (see SERVER-43074), so the safest option is to report a validation
error, which requires rebuilding the index to fix the problem.