SERVER-7694 Enable allowing disk use for sorts in the find command.
For find commands that request a sort, the query execution
engine can now spill data to disk if the memory requirements
would exceed 'internalQueryExecBlockingSortBytes' and the
'allowDiskUse' parameter is set to true. By default, the
memory threshold is currently 100MB. This allows
applications which need to sort server-side to no longer be
subject to an arbitrary data size threshold.