SERVER-44608 Only set "use44SortKeys" in command from mongos.
In a mixed cluster, it is possible for a 4.4 shard to create a command
that will get sent to a 4.2 shard. If the command includes the
"use44SortKeys" flag, it will fail on 4.2 mongod.

No shard-to-shard pipeline will ever include a change stream, so
"use44SortKeys" is unnecessary in this case.